+++ utils/utils_noise.py
@@ -1,20 +1,5 @@
-import cv2
 import numpy as np
 
-def save_img(picture, path_picture):
-    picture_temp = picture
-    if(np.ndim(picture) == 3):
-        # color
-        picture_temp = np.moveaxis(picture, 0, 2)
-    picture_temp[picture_temp > 1.] = 1.
-    picture_temp[picture_temp < 0.] = 0.
-
-    cv2.imwrite(path_picture, np.uint8(picture_temp*255.))
-    
-def save_imgs(pictures, path_pictures, format):
-    for i, picture in enumerate(pictures):
-        save_img(picture = picture, path_picture = path_pictures[i] + format)
-        
 def add_salt_and_pepper_noise(img, noise_level, random_sampling_op):
     noise_cnt = (int)(img.shape[-2] * img.shape[-1] * noise_level / 2)
     noise_target = np.ones([img.shape[-2], img.shape[-1]])
